move game-logic to appropriate module
